pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "numexpr>=2.14.1",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.22",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "black>=24.0.0",
]
